+++ R/PatientProfiles-package.R
@@ -18,6 +18,7 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom lifecycle deprecated
 #' @importFrom rlang %||%
 #' @importFrom rlang .data
 #' @importFrom rlang .env